requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
baostock>=0.8.0
openai>=1.0.0
python-dotenv>=1.0.0
ta>=0.10.0
plotly>=5.18.0
pyinstaller>=6.0.0
pywebview>=4.0.0
requests>=2.31.0
toml>=0.10.0
watchdog>=3.0.0
packaging>=23.0
filelock>=3.13.0
